drf-intro/simple_crud/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.8.6
pillow==8.4.0
